+++ packages/app/src/cli/services/dev/select-app.ts
@@ -30,8 +30,8 @@ export async function selectOrCreateApp(
   if (createNewApp) {
     return createApp(org, localAppName, token)
   } else {
-    const selectedApp = await selectAppPrompt(apps, org.id, token)
-    const fullSelectedApp = await fetchAppFromApiKey(selectedApp.apiKey, token)
+    const selectedAppApiKey = await selectAppPrompt(apps, org.id, token)
+    const fullSelectedApp = await fetchAppFromApiKey(selectedAppApiKey, token)
     return fullSelectedApp!
   }
 }
@@ -55,7 +55,6 @@ export async function createApp(org: Organization, appName: string, token: strin
     throw new error.Abort(errors)
   }
 
-  output.success(`${result.appCreate.app.title} has been created on your Partners account`)
   const createdApp: OrganizationApp = result.appCreate.app
   createdApp.organizationId = org.id
   createdApp.newApp = true