+++ packages/cli-kit/src/private/node/ui/components/TextInput.tsx
@@ -6,19 +6,30 @@ import chalk from 'chalk'
 import type {FC} from 'react'
 
 interface Props {
-  placeholder?: string
+  defaultValue?: string
   value: string
   onChange: (value: string) => void
   color?: string
+  password?: boolean
+  focus?: boolean
+  placeholder?: string
 }
 
-const TextInput: FC<Props> = ({value, placeholder = '', onChange, color = 'cyan'}) => {
-  const [cursorOffset, setCursorOffset] = useState((value || '').length)
+const TextInput: FC<Props> = ({
+  value: originalValue,
+  defaultValue = '',
+  onChange,
+  placeholder = '',
+  color = 'cyan',
+  password = false,
+  focus = true,
+}) => {
+  const [cursorOffset, setCursorOffset] = useState((originalValue || '').length)
 
   // if the updated value is shorter than the last one we need to reset the cursor
   useEffect(() => {
     setCursorOffset((previousOffset) => {
-      const newValue = value || ''
+      const newValue = originalValue || ''
 
       if (previousOffset > newValue.length - 1) {
         return newValue.length
@@ -26,11 +37,21 @@ const TextInput: FC<Props> = ({value, placeholder = '', onChange, color = 'cyan'
 
       return previousOffset
     })
-  }, [value])
+  }, [originalValue])
 
-  let renderedValue = value.length > 0 ? '' : chalk.inverse(' ')
+  const value = password ? '*'.repeat(originalValue.length) : originalValue
+  let renderedValue
+
+  const renderPlaceholder = (value: string) => {
+    return chalk.inverse(value[0]) + chalk.dim(value.slice(1))
+  }
+
   const renderedPlaceholder =
-    placeholder.length > 0 ? chalk.inverse(placeholder[0]) + chalk.dim(placeholder.slice(1)) : undefined
+    defaultValue.length > 0
+      ? renderPlaceholder(defaultValue)
+      : placeholder.length > 0
+      ? renderPlaceholder(placeholder)
+      : chalk.inverse(' ')
 
   // render cursor
   renderedValue = value
@@ -48,49 +69,51 @@ const TextInput: FC<Props> = ({value, placeholder = '', onChange, color = 'cyan'
     renderedValue += chalk.inverse(' ')
   }
 
-  useInput((input, key) => {
-    if (
-      key.upArrow ||
-      key.downArrow ||
-      (key.ctrl && input === 'c') ||
-      key.tab ||
-      (key.shift && key.tab) ||
-      key.return
-    ) {
-      return
-    }
-
-    let nextCursorOffset = cursorOffset
-    let nextValue = value
-
-    if (key.leftArrow) {
-      if (cursorOffset > 0) {
-        nextCursorOffset--
+  useInput(
+    (input, key) => {
+      if (
+        key.upArrow ||
+        key.downArrow ||
+        (key.ctrl && input === 'c') ||
+        key.tab ||
+        (key.shift && key.tab) ||
+        key.return
+      ) {
+        return
       }
-    } else if (key.rightArrow) {
-      if (cursorOffset < value.length) {
-        nextCursorOffset++
+
+      let nextCursorOffset = cursorOffset
+      let nextValue = originalValue
+
+      if (key.leftArrow) {
+        if (cursorOffset > 0) {
+          nextCursorOffset--
+        }
+      } else if (key.rightArrow) {
+        if (cursorOffset < originalValue.length) {
+          nextCursorOffset++
+        }
+      } else if (key.backspace || key.delete) {
+        if (cursorOffset > 0) {
+          nextValue = originalValue.slice(0, cursorOffset - 1) + originalValue.slice(cursorOffset, originalValue.length)
+          nextCursorOffset--
+        }
+      } else {
+        nextValue =
+          originalValue.slice(0, cursorOffset) + input + originalValue.slice(cursorOffset, originalValue.length)
+        nextCursorOffset += input.length
       }
-    } else if (key.backspace || key.delete) {
-      if (cursorOffset > 0) {
-        nextValue = value.slice(0, cursorOffset - 1) + value.slice(cursorOffset, value.length)
-        nextCursorOffset--
+
+      setCursorOffset(nextCursorOffset)
+
+      if (nextValue !== originalValue) {
+        onChange(nextValue)
       }
-    } else {
-      nextValue = value.slice(0, cursorOffset) + input + value.slice(cursorOffset, value.length)
-      nextCursorOffset += input.length
-    }
-
-    setCursorOffset(nextCursorOffset)
-
-    if (nextValue !== value) {
-      onChange(nextValue)
-    }
-  })
-
-  return (
-    <Text color={color}>{placeholder ? (value.length > 0 ? renderedValue : renderedPlaceholder) : renderedValue}</Text>
+    },
+    {isActive: focus},
   )
+
+  return <Text color={color}>{value.length > 0 ? renderedValue : renderedPlaceholder}</Text>
 }
 
 export {TextInput}